Adding the training code for adversarial coreset selection
Added the training code for adversarial coreset selection. Currently, the code is repetitive and the main trainer seems long. To get rid of this, we need to pay close attention to the TODOs in the uploaded files and address them one by one. Nevertheless, the current code should run just fine.